performance_notifier: move Query & Route to separate files
These structs are not tightly coupled with PerformanceNotifier. It makes sense
to have them in separate files to reduce the churn of PerformanceNotifier.
