+++ lib/airbrake-ruby.rb
@@ -42,6 +42,8 @@ require 'airbrake-ruby/deploy_notifier'
 require 'airbrake-ruby/stat'
 require 'airbrake-ruby/time_truncate'
 require 'airbrake-ruby/tdigest'
+require 'airbrake-ruby/query'
+require 'airbrake-ruby/route'
 
 # This module defines the Airbrake API. The user is meant to interact with
 # Airbrake via its public class methods. Before using the library, you must to
